Extract MyTardis object and endpoint information into separate modules, and move a MyTardis-related helper from overseer to the MyTardis client
