+++ src/conveyor/conveyor.py
@@ -8,7 +8,7 @@ from tempfile import NamedTemporaryFile
 
 from src.blueprints.datafile import Datafile, DatafileReplica
 from src.config.config import FilesystemStorageBoxConfig
-from src.overseers import helpers as overseer
+from src.mytardis_client import helpers as mt_helpers
 
 logger = logging.getLogger(__name__)
 
@@ -56,7 +56,7 @@ class Conveyor:
             DatafileReplica: The Replica representing the copied file.
         """
         file_path = file.directory / file.filename
-        dataset_id = overseer.resource_uri_to_id(file.dataset)
+        dataset_id = mt_helpers.resource_uri_to_id(file.dataset)
         return DatafileReplica(
             protocol="file",
             location=self._store.storage_name,
@@ -116,7 +116,7 @@ class Conveyor:
         files_by_dataset: dict[int, list[Datafile]] = {}
         for df in dfs:
             # Group datafiles by the dataset they are in.
-            dataset_id = overseer.resource_uri_to_id(df.dataset)
+            dataset_id = mt_helpers.resource_uri_to_id(df.dataset)
             if dataset_id not in files_by_dataset:
                 files_by_dataset[dataset_id] = []
             files_by_dataset[dataset_id].append(df)
