+++ src/overseers/object_matchers.py
@@ -2,8 +2,8 @@
 
 from typing import Any
 
-from src.mytardis_client.enumerators import MyTardisObjectType
-from src.overseers.helpers import resource_uri_to_id
+from src.mytardis_client.helpers import resource_uri_to_id
+from src.mytardis_client.types import MyTardisObjectType
 
 MatchKeys = dict[str, Any]
